refactor : custom text field, custom dropdown menu view, controller 분리
