+++ presentation/src/main/java/kgb/plum/presentation/ui/common/textField/CustomTextField.kt
@@ -1,22 +1,17 @@
-package kgb.plum.presentation.ui.common
+package kgb.plum.presentation.ui.common.textField
 
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.text.BasicTextField
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.Title
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
@@ -56,18 +51,5 @@ fun CustomTextField(
       ),
       contentPadding = PaddingValues(0.dp, 0.dp, 10.dp, 0.dp), // 패딩 삭제
     )
-  }
-}
-
-class CustomTextFieldController(private val onTextChangeCallback: () -> Unit = {}) {
-  var text by mutableStateOf("")
-  fun onTextChange(value: String, callback: () -> Unit = onTextChangeCallback) {
-    text = value
-    println("CustomTexField : $text, $value")
-    callback()
-  }
-
-  fun clearText() {
-    text = ""
   }
 }
