+++ presentation/src/main/java/kgb/plum/presentation/ui/components/recruit/RecruitHeader.kt
@@ -4,7 +4,6 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.material3.ButtonDefaults
@@ -16,8 +15,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import kgb.plum.presentation.model.SortType
-import kgb.plum.presentation.ui.common.CustomDropdownMenuController
-import kgb.plum.presentation.ui.common.CustomTextDropdownMenu
+import kgb.plum.presentation.ui.common.dropdown.CustomDropdownMenuController
+import kgb.plum.presentation.ui.common.dropdown.CustomTextDropdownMenu
 import kgb.plum.presentation.ui.theme.Padding
 
 @Composable
