+++ presentation/src/main/java/kgb/plum/presentation/viewmodel/EditListScreenViewModel.kt
@@ -2,7 +2,7 @@ package kgb.plum.presentation.viewmodel
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kgb.plum.presentation.ui.common.CustomTextFieldController
+import kgb.plum.presentation.ui.common.textField.CustomTextFieldController
 import javax.inject.Inject
 
 @HiltViewModel
